tags: allow non-alphanumeric characters
Tag name is filtered using slugify default mode (filtering out all spaces and non-alphanumeric characters).
It causes tags like 'c' and 'c++' to be merged together and multiple index file for the same tag are generated.

To fix it set slugify mode to 'pretty' so some special characters (including '+') are allowed.
Jekyll-archives plugins should also be updated to the version 2.3, so the 'slug_mode' configuration parameter can be set to 'pretty'.